chore: incorporate Group component into current structure
